% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{epi}
\alias{epi}
\title{A spatial S-E-I-R dataset}
\format{\code{epi} is a dataframe with 150 rows with 8 columns named \code{k}, \code{coor_x}, \code{coor_y}, \code{te}, \code{ti}, \code{tr}, \code{stb}, \code{infected_source}. Columns correspond to indices for individuals, x-coordinate of individuals in space, y-coordinate, exposure times, times become infectious, recovery/removal times, suitability/susceptibility of sites, sources of infections.}
\usage{
epi
}
\description{
An epidemic simulated from a spatial S-E-I-R model.
}
\author{
Max S.Y. Lau
}

